belica-bot/requirements.txt
# Discord Bot
discord.py>=2.3.0

# HTTP client for GraphQL API
aiohttp>=3.9.0

# Environment variable management
python-dotenv>=1.0.0

